Address Copilot review: fix history timing, cancellation, dedup post-processing, computed opacity
- Record history after post-processing so restored clips include all processing
- Wrap ApplyPostProcessing in OperationCanceledException handling with cancellation checks between steps
- Extract shared TtsPostProcessor helper used by both TextToSpeechViewModel and ReviewSpeechViewModel
- Make HistoryButtonOpacity a computed property derived from HasHistory
